add PR to CHANGES.md move `EVT_VARIANT` and `EVT_PRC_LOGIN` to the outside change Pointer to WSing in `EVT_RPC_LOGIN` add TypeMapper to `EVT_VARIANT` and `EVT_PRC_LOGIN`
